+++ src/python/DAS/services/combined/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python
+#-*- coding: ISO-8859-1 -*-
+
+"""
+DAS combined service module
+"""
+__author__ = "Valentin Kuznetsov"